update user policy and seeders

+++ app/Http/Controllers/SiteController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Constants\CurrencyType;
+use App\Constants\PermissionSlug;
+use App\Constants\PolicyName;
 use App\Constants\TypeName;
 use App\Http\Requests\Site\StoreRequest;
 use App\Http\Requests\Site\UpdateRequest;
@@ -13,11 +15,13 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 use Inertia\Response;
+use Spatie\Permission\Models\Role;
 
 class SiteController extends Controller
 {
     public function index(): Response
     {
+
         $sites = Site::with('type')->where('user_id', Auth::id())->get();
 
         return Inertia::render('Site/Index', compact('sites'));
@@ -25,6 +29,10 @@ class SiteController extends Controller
 
     public function create(): Response
     {
+        if(!Auth::user()->can(PermissionSlug::SITES_CREATE)){
+            abort(403);
+        }
+
         $types = TypeName::toArray();
         $currencies = CurrencyType::toArray();
 
@@ -37,6 +45,7 @@ class SiteController extends Controller
 
     public function store(StoreRequest $request): RedirectResponse
     {
+
         if (! Auth::check()) {
 
             return redirect()->route('avatar');
@@ -58,6 +67,7 @@ class SiteController extends Controller
 
     public function show(Site $site): Response
     {
+
         $types = TypeName::toArray();
         $currencies = CurrencyType::toArray();
 
@@ -73,6 +83,9 @@ class SiteController extends Controller
      */
     public function edit(Site $site): Response
     {
+        if(!Auth::user()->can(PermissionSlug::SITES_UPDATE)){
+            abort(403);
+        }
         $types = TypeName::toArray();
         $currencies = CurrencyType::toArray();
 
@@ -88,7 +101,9 @@ class SiteController extends Controller
      */
     public function update(UpdateRequest $request, Site $site): RedirectResponse
     {
-
+        if(!Auth::user()->can(PermissionSlug::SITES_UPDATE)){
+            abort(403);
+        }
         $data = $request->except('avatar');
 
         if ($request->hasFile('avatar')) {
@@ -112,6 +127,9 @@ class SiteController extends Controller
      */
     public function destroy(Site $site): RedirectResponse
     {
+        if(!Auth::user()->can(PermissionSlug::SITES_DELETE)){
+            abort(403);
+        }
         if ($site->avatar) {
             Storage::disk('public')->delete($site->avatar);
         }
