+++ app/Http/Controllers/UserController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\PermissionSlug;
+use App\Constants\PolicyName;
 use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Inertia\Response;
 use Spatie\Permission\Models\Role;
@@ -13,6 +16,10 @@ class UserController extends Controller
 {
     public function index(): Response
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_VIEW_ANY)){
+            ABORT(403);
+        }
+
         $users = User::with('roles')->get();
 
         return inertia('User/Index', [
@@ -22,11 +29,17 @@ class UserController extends Controller
 
     public function create(): Response
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_CREATE)){
+            abort(403);
+        }
         return inertia('User/Create');
     }
 
     public function store(StoreUserRequest $request): \Illuminate\Http\RedirectResponse
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_CREATE)){
+            abort(403);
+        }
         $user = new User();
         $user->name = $request->name;
         $user->email = $request->email;
@@ -37,6 +50,9 @@ class UserController extends Controller
     }
     public function show(User $user): Response
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_VIEW)){
+            abort(403);
+        }
         $roles = $user->roles()->get();
 
         return inertia('User/Show', [
@@ -47,6 +63,9 @@ class UserController extends Controller
 
     public function edit(User $user): Response
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_UPDATE)){
+            abort(403);
+        }
         $roles = Role::all();
         return inertia('User/Edit', [
             'user' => [
@@ -60,7 +79,9 @@ class UserController extends Controller
 
     public function update(UpdateUserRequest $request, User $user): \Illuminate\Http\RedirectResponse
     {
-
+        if(!Auth::user()->can(PermissionSlug::USERS_UPDATE)){
+            abort(403);
+        }
         $user->update([
             'name' => $request->input('name'),
         ]);
@@ -72,6 +93,9 @@ class UserController extends Controller
 
     public function destroy(User $user): Response
     {
+        if(!Auth::user()->can(PermissionSlug::USERS_DELETE)){
+            abort(403);
+        }
         $user->delete();
 
         return inertia('User/Index');
